requirements.txt
pydantic>=2.0
google-genai>=1.0
groq>=0.9
PyMuPDF>=1.24
pytesseract>=0.3
Pillow>=10.0
fastapi>=0.115
uvicorn>=0.30
jinja2>=3.1
weasyprint>=62
xhtml2pdf>=0.2.11
python-dotenv>=1.0
python-multipart>=0.0.9
